4129: Extract chat fab
