+++ web/src/components/ChatContainer.tsx
@@ -1,7 +1,4 @@
-import QuestionAnswerOutlinedIcon from '@mui/icons-material/QuestionAnswerOutlined'
 import { dialogContentClasses } from '@mui/material/DialogContent'
-import Fab from '@mui/material/Fab'
-import Typography from '@mui/material/Typography'
 import { styled } from '@mui/material/styles'
 import React, { ReactElement, useContext } from 'react'
 import { useSearchParams } from 'react-router'
@@ -16,22 +13,11 @@ import useLocalStorage from '../hooks/useLocalStorage'
 import useLockedBody from '../hooks/useLockedBody'
 import { chatIdKey, generateChatId } from '../utils/chat'
 import ChatController from './ChatController'
+import ChatFab from './ChatFab'
 import ChatMenu from './ChatMenu'
 import HeaderLanguageSelectorItem from './HeaderLanguageSelectorItem'
 import { LanguageChangePath } from './LanguageList'
 import Dialog from './base/Dialog'
-
-const ChatButtonContainer = styled('div')<{ bottom: number }>`
-  position: fixed;
-  bottom: ${props => props.bottom}px;
-  inset-inline-end: 16px;
-  margin-bottom: 16px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: min-content;
-  gap: 8px;
-`
 
 const StyledDialog = styled(Dialog)({
   [`.${dialogContentClasses.root}`]: {
@@ -48,46 +34,45 @@ type ChatContainerProps = {
 const ChatContainer = ({ region, languageCode, languageChangePaths }: ChatContainerProps): ReactElement | null => {
   const [queryParams, setQueryParams] = useSearchParams()
   const chatVisible = parseQueryParams(queryParams).chat ?? false
-  const { desktop, xsmall, visibleFooterHeight, bottomNavigationHeight } = useDimensions()
+  const { xsmall } = useDimensions()
   const { visible: ttsPlayerVisible } = useContext(TtsContext)
+  useLockedBody(chatVisible)
 
   const [chatId, setChatId] = useLocalStorage<string>({
     key: chatIdKey(region.code),
     initialValue: generateChatId(),
   })
 
-  const chatName = getChatName(buildConfig().appName)
-  useLockedBody(chatVisible)
-
-  const hideChatButton = xsmall && ttsPlayerVisible
-
-  const open = () => {
+  const openChat = () => {
     const newQueryParams = queryParams
     newQueryParams.set(CHAT_QUERY_KEY, 'true')
     setQueryParams(newQueryParams)
   }
 
-  const close = () => {
+  const closeChat = () => {
     const newQueryParams = queryParams
     newQueryParams.delete(CHAT_QUERY_KEY)
     setQueryParams(newQueryParams)
   }
 
+  const hideChatButton = xsmall && ttsPlayerVisible
   if (hideChatButton) {
     return null
   }
 
   if (chatVisible) {
+    const chatName = getChatName(buildConfig().appName)
     const chatQuery = toQueryParams({ chat: true }).toString()
     const chatLanguageChangePaths =
       languageChangePaths?.map(({ path, ...rest }) => ({
         ...rest,
         path: path ? `${path}?${chatQuery}` : null,
       })) ?? []
+
     return (
       <StyledDialog
         title={chatName}
-        close={close}
+        close={closeChat}
         actions={[
           ...(languageChangePaths
             ? [
@@ -105,18 +90,7 @@ const ChatContainer = ({ region, languageCode, languageChangePaths }: ChatContai
     )
   }
 
-  return (
-    <ChatButtonContainer bottom={bottomNavigationHeight ?? visibleFooterHeight}>
-      <Fab onClick={open} color='primary' aria-label={chatName}>
-        <QuestionAnswerOutlinedIcon fontSize='large' />
-      </Fab>
-      {desktop && (
-        <Typography textAlign='center' aria-hidden>
-          {chatName}
-        </Typography>
-      )}
-    </ChatButtonContainer>
-  )
+  return <ChatFab onClick={openChat} />
 }
 
 export default ChatContainer
