+++ web/src/components/ChatFab.tsx
@@ -1,0 +1,46 @@
+import QuestionAnswerOutlinedIcon from '@mui/icons-material/QuestionAnswerOutlined'
+import Fab from '@mui/material/Fab'
+import Typography from '@mui/material/Typography'
+import { styled } from '@mui/material/styles'
+import React, { ReactElement } from 'react'
+
+import { getChatName } from 'shared'
+
+import buildConfig from '../constants/buildConfig'
+import useDimensions from '../hooks/useDimensions'
+
+const ChatButtonContainer = styled('div')<{ bottom: number }>`
+  position: fixed;
+  bottom: ${props => props.bottom}px;
+  inset-inline-end: 16px;
+  margin-bottom: 16px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  width: min-content;
+  gap: 8px;
+`
+
+type ChatButtonProps = {
+  onClick: () => void
+}
+
+const ChatFab = ({ onClick }: ChatButtonProps): ReactElement => {
+  const { desktop, visibleFooterHeight, bottomNavigationHeight } = useDimensions()
+  const chatName = getChatName(buildConfig().appName)
+
+  return (
+    <ChatButtonContainer bottom={bottomNavigationHeight ?? visibleFooterHeight}>
+      <Fab onClick={onClick} color='primary' aria-label={chatName}>
+        <QuestionAnswerOutlinedIcon fontSize='large' />
+      </Fab>
+      {desktop && (
+        <Typography textAlign='center' aria-hidden>
+          {chatName}
+        </Typography>
+      )}
+    </ChatButtonContainer>
+  )
+}
+
+export default ChatFab
